+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for small-bytearray-builder
 
+## 0.3.3.0 -- 2020-??-??
+
+* Add `word64PaddedLowerHex` and `word32PaddedLowerHex`
+* Add `word{128,256}Padded{Lower,Upper}Hex`
+
 ## 0.3.2.0 -- 2020-01-20
 
 * Add `putMany`, which allows pasting into the same mutable byte